Forced the database from mysql to postgresql
- Changed backup manager methods to use pg_dump and pg_restore for PostgreSQL schema management.
- Updated database models to enforce non-null constraints and added unique constraints where necessary.
- Enhanced daily stats retrieval to validate user and language existence before processing.
- Modified various model fields to ensure proper data types and defaults.
- Improved media cleanup tasks to handle orphaned files more effectively.
- Added error handling and logging improvements across services.

+++ src/lapp/models/data_collection/commitment_log.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Boolean, String
+from sqlalchemy import Column, Integer, Boolean, String, Float
 
 from ..base import BaseDataCollectionModel
 
@@ -7,7 +7,7 @@ class CommitmentLog(BaseDataCollectionModel):
 
     days_active = Column(Integer, nullable=False)
     total_items_reviewed = Column(Integer, nullable=False)
-    total_time_ms = Column(Integer, nullable=False)
+    total_time_ms = Column(Float, nullable=False)
     longest_streak_ever = Column(Integer, nullable=False)
     streak_last_computed_at = Column(String, nullable=False)  # ISO datetime string
